requirements.txt
Django==4.2.10
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
httpx==0.28.1
python-decouple==3.8
python-dotenv==1.1.1
rest_framework_simplejwt==0.0.2
